+++ CHANGELOG.md
@@ -6,6 +6,11 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased] - ReleaseDate
 ### Added
 
+- Added the ability to mock generic structs with generic methods whose only
+  generic types are lifetimes.  This is useful for mocking generic structs that
+  implement traits like `Future` and `Stream`.
+  ([#226](https://github.com/asomers/mockall/pull/226))
+
 - Added the ability to mock methods returning references to trait objects.
   ([#213](https://github.com/asomers/mockall/pull/213))